Add packages and refactor code accordingly
