--- src/main/java/ExitCommand.java
@@ -1,8 +1,0 @@
-public class ExitCommand extends Command{
-    public ExitCommand() {}
-
-    @Override
-    public boolean isExit() {
-        return true;
-    }
-}
