+++ src/main/java/garfield/commands/Command.java
@@ -1,3 +1,10 @@
+package garfield.commands;
+
+import garfield.exceptions.GarfieldException;
+import garfield.storage.Storage;
+import garfield.tasks.TaskList;
+import garfield.ui.Ui;
+
 public abstract class Command {
     public Command() {}
     public boolean isExit() {
