+++ src/main/java/garfield/commands/ExitCommand.java
@@ -1,0 +1,12 @@
+package garfield.commands;
+
+import garfield.commands.Command;
+
+public class ExitCommand extends Command {
+    public ExitCommand() {}
+
+    @Override
+    public boolean isExit() {
+        return true;
+    }
+}
