+++ src/main/java/garfield/commands/MarkCommand.java
@@ -1,3 +1,10 @@
+package garfield.commands;
+
+import garfield.exceptions.GarfieldException;
+import garfield.storage.Storage;
+import garfield.tasks.TaskList;
+import garfield.ui.Ui;
+
 public class MarkCommand extends Command {
     private int taskId;
 
@@ -7,7 +14,7 @@ public class MarkCommand extends Command {
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) throws GarfieldException {
-        ui.showMessage("Nice. You actually did something. I've marked that task as done:\n\n\t"
+        ui.showMessage("Nice. You actually did something. I've marked that garfield.task as done:\n\n\t"
                 + taskList.mark(taskId));
         storage.save(taskList);
