+++ src/main/java/garfield/exceptions/GarfieldException.java
@@ -1,3 +1,5 @@
+package garfield.exceptions;
+
 public class GarfieldException extends Exception {
     public GarfieldException(String message) {
         super(message);
