+++ src/main/java/garfield/tasks/Event.java
@@ -1,3 +1,5 @@
+package garfield.tasks;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
