+++ text-ui-test/input.txt
@@ -1,6 +1,6 @@
-todo task 1
-deadline task 2 /by Sunday
-event task 3 /from Today /to next week
+todo garfield.tasks 1
+deadline garfield.tasks 2 /by Sunday
+event garfield.tasks 3 /from Today /to next week
 mark 1
 unmark 1
 mark 2
